important logic fix: entropy calculation includes probabilities. converging issues are over.

+++ HarmonyDB.Index/HarmonyDB.Index.Analysis.Em/Models/Constants.cs
@@ -5,8 +5,8 @@ public static class Constants
     public const int TonicCount = 12; // Number of different tonics
     public const int ScaleCount = 2;  // Major and Minor
 
-    public static IEnumerable<(byte tonic, Scale scale)> Pairs
-        => Indices.Select(x => ((byte)x.tonic, (Scale)x.scale));
+    public static IEnumerable<(int tonic, Scale scale)> Pairs
+        => Indices.Select(x => (x.tonic, (Scale)x.scale));
 
     public static IEnumerable<(int tonic, int scale)> Indices
         => Enumerable
